+++ R/f7Popup.R
@@ -2,8 +2,13 @@
 #'
 #' @param ... Content.
 #' @param id Popup unique id.
-#' @param label Popup trigger label.
 #' @param title Title.
+#' @param backdrop Enables Popup backdrop (dark semi transparent layer behind). Default to TRUE.
+#' @param closeByBackdropClick When enabled, popup will be closed on backdrop click. Default to TRUE.
+#' @param closeOnEscape When enabled, popup will be closed on ESC keyboard key press. Default to FALSE.
+#' @param animate Whether the Popup should be opened/closed with animation or not. Default to TRUE.
+#' @param swipeToClose Whether the Popup can be closed with swipe gesture. Can be true to allow to close popup with swipes to top and to bottom.
+#' Default to FALSE.
 #'
 #' @export
 #'
@@ -21,43 +26,76 @@
 #'         hairline = FALSE,
 #'         shadow = TRUE
 #'       ),
+#'       f7Button("togglePopup", "Toggle Popup"),
 #'       f7Popup(
 #'        id = "popup1",
-#'        label = "Open",
 #'        title = "My first popup",
-#'        "Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-#'          Quisque ac diam ac quam euismod porta vel a nunc. Quisque sodales
-#'          scelerisque est, at porta justo cursus ac"
+#'        f7Text("text", "Popup content", "This is my first popup ever, I swear!"),
+#'        verbatimTextOutput("popupContent")
 #'       )
 #'      )
 #'    ),
-#'    server = function(input, output) {}
+#'    server = function(input, output, session) {
+#'
+#'     output$popupContent <- renderPrint(input$text)
+#'
+#'     observeEvent(input$togglePopup, {
+#'      f7TogglePopup(id = "popup1")
+#'     })
+#'
+#'     observeEvent(input$popup1, {
+#'
+#'      popupStatus <- if (input$popup1) "opened" else "closed"
+#'
+#'      f7Toast(
+#'       session,
+#'       position = "top",
+#'       text = paste("Popup is", popupStatus)
+#'      )
+#'     })
+#'    }
 #'  )
 #' }
-f7Popup <- function(..., id, label = "Open", title) {
+f7Popup <- function(..., id, title, backdrop = TRUE, closeByBackdropClick = TRUE,
+                    closeOnEscape = FALSE, animate = TRUE, swipeToClose = FALSE) {
 
-  shiny::tagList(
-    shiny::tags$div(
-      class = "block",
-      # button handler (maybe twick f7Buttons to be able to trigger a popup)
-      shiny::a(
-        class = "link popup-open",
-        href = "#",
-        `data-popup` = paste0("#", id),
-        label
-      )
-    ),
-    shiny::tags$div(
+  popupProps <- dropNulls(
+    list(
       class = "popup popup-tablet-fullscreen",
       id = id,
-      shiny::br(),
-      shiny::br(),
-      shiny::div(
-        class = "block",
-        shiny::p(title),
-        shiny::p(shiny::a(class = "link popup-close", href = "#", "Close")),
-        shiny::p(...)
-      )
+      `data-backdrop` = tolower(backdrop),
+      `data-close-by-backdrop-click` = tolower(closeByBackdropClick),
+      `data-close-on-escape` = tolower(closeOnEscape),
+      `data-animate` = tolower(animate),
+      `data-swipe-to-close` = tolower(swipeToClose)
     )
   )
+
+  popupTag <- do.call(shiny::tags$div, popupProps)
+  popupWrap <- shiny::tagAppendChildren(
+    popupTag,
+    f7InputsDeps(),
+    shiny::br(),
+    shiny::br(),
+    shiny::div(
+      class = "block",
+      shiny::p(title),
+      shiny::p(shiny::a(class = "link popup-close", href = "#", "Close")),
+      shiny::p(...)
+    )
+  )
+
+  popupWrap
+}
+
+
+
+#' Toggle \link{f7Popup}.
+#'
+#' @param id Popup id.
+#' @param session Shiny session.
+
+#' @export
+f7TogglePopup <- function(id, session = shiny::getDefaultReactiveDomain()) {
+  session$sendInputMessage(id, message = NULL)
 }